Store Messages at Server End
Potential for a database implementation, but I have created a basic proof of this python MQTT listener that could receive payloads from the Arduino and package them with relevant metadata, finally pushing them to a database that a web portal can pull attendance from. For now, we hold data in the running instance.